requirements.txt
fastapi
uvicorn
groq
tavily-python
httpx
python-dotenv
pydantic
